PATCH: [ 2209911 ] Update PCRE to version 7.8  Update to latest PCRE version available from http://www.pcre.org/

+++ Externals/pcre/pcregrep.c
@@ -1054,10 +1054,10 @@ while (ptr < endptr)
         if (printname != NULL) fprintf(stdout, "%s:", printname);
         if (number) fprintf(stdout, "%d:", linenumber);
         if (line_offsets)
-          fprintf(stdout, "%d,%d", matchptr + offsets[0] - ptr,
+          fprintf(stdout, "%d,%d", (int)(matchptr + offsets[0] - ptr),
             offsets[1] - offsets[0]);
         else if (file_offsets)
-          fprintf(stdout, "%d,%d", filepos + matchptr + offsets[0] - ptr,
+          fprintf(stdout, "%d,%d", (int)(filepos + matchptr + offsets[0] - ptr),
             offsets[1] - offsets[0]);
         else
           fwrite(matchptr + offsets[0], 1, offsets[1] - offsets[0], stdout);
